JWT authentication has been added for the users and is attached as a
middleware. Tokens are returned when users are created or logged in.
These services have been implemented but CRUD operations still need
to be added.

Password hashing has been added using .NET solutions. And we can
successfully create users and log them in.

The main program file has also been reorganized to make it easier to
read and understand. We had a lot of configuration and services being
added which was messy. This has been abstracted to a Helpers.cs file.