Updated README.MD
Environment variables are preferred over using the App settings.
I find it easier to use environment variables when deploying to Azure,
since these will have to be set in the Azure portal anyway.

I have changed the README.MD to reflect the changes. This commit also
closes issue #3, the remaining CRUD operations for the User model have
been added.

At a later time we can configure a password reset option, for now
the configured put operation will not allow the password to be changed.