Updated authentication methods to use a basic token as a first step.
If the token is valid then we will authenticate the user in subsequent
requests using their JWT. If not we will return a 401.

After some thinking, the WebApp model is irrelevant and has been removed
from the project.

The README has been updated to reflect the configuration changes.

The DotEnv class has been updated to stop splitting the string after the
first equals sign.